Added exception when comparing non valid types
Doing "10" >= 5 will raise an Error just like the original Lua
Interpreter http://codepad.org/0cLOuQFS . Previously when for example an
Lua integer and Lua string object was compared. The statement was always
false. This can be very dangerous. Now the propper "attempt to compare X
with Y" error is raised.